+++ src/Codifier.php
@@ -16,23 +16,11 @@ class Codifier {
     protected static $hidden_labels = [];
 
     /**
-     * This is used with WP object caching.
-     * We use ACF's cache group for now.
+     * This array holds all plugin settings which are overridable with constants.
      *
-     * This is overridable with the 'ACF_CODIFIER_CACHE_GROUP' constant.
-     *
-     * @var string
+     * @var array
      */
-    public static $cache_group = 'acf';
-
-    /**
-     * This ttl is used when objects are not eternally cached.
-     *
-     * This is overridable with the 'ACF_CODIFIER_CACHE_TTL' constant.
-     *
-     * @var float|int
-     */
-    public static $cache_ttl = 15 * MINUTE_IN_SECONDS;
+    protected static $settings = [];
 
     /**
      * Initialize the plugin by adding hooks.
@@ -41,29 +29,75 @@ class Codifier {
      */
     public static function init() {
 
-        self::override_settings();
+        self::load_settings();
 
         add_action( 'admin_head', __CLASS__ . '::hide_labels' );
 
         // Cache control.
-        add_action( 'updated_postmeta', [ __CLASS__, 'flush_post_meta_caches' ], 1, 1 );
-        add_filter( 'acf/update_value',  [ __CLASS__, 'flush_formatted_value' ], 10, 3 );
+        add_action( 'updated_postmeta', [ __CLASS__, 'flush_meta_cache' ], 1, 3 );
 
     }
 
     /**
-     * This function overrides class settings with defined constants.
+     * This function loads the plugin settings and possibly overrides them with constants.
      */
-    public static function override_settings(  ) {
+    protected static function load_settings() {
 
-        if ( defined( 'ACF_CODIFIER_CACHE_GROUP' ) ) {
-            self::$cache_group = ACF_CODIFIER_CACHE_GROUP;
+        /**
+         * This is the WP object cache group key for field objects.
+         *
+         * This is overridable with the 'ACF_CODIFIER_FIELD_OBJECT_GROUP' constant.
+         *
+         * @var string Defaults to 'acf_codifier/field_object'.
+         */
+        self::$settings['field_object_group'] = defined( 'ACF_CODIFIER_FIELD_OBJECT_GROUP' ) ?
+            ACF_CODIFIER_FIELD_OBJECT_GROUP : 'acf_codifier/field_object';
+
+        /**
+         * This is the WP object cache group key for objects' sorted metadata.
+         *
+         * This is overridable with the 'ACF_CODIFIER_SORTED_META_GROUP' constant.
+         *
+         * @var string Defaults to 'acf_codifier/sorted_meta'.
+         */
+        self::$settings['sorted_meta_group'] = defined( 'ACF_CODIFIER_SORTED_META_GROUP' ) ?
+            ACF_CODIFIER_SORTED_META_GROUP : 'acf_codifier/sorted_meta';
+
+        /**
+         * This is the WP object cache group key loaded field values.
+         *
+         * This is overridable with the 'ACF_CODIFIER_FIELD_VALUE' constant.
+         *
+         * @var string Defaults to 'acf_codifier/field_value'.
+         */
+        self::$settings['field_value_group'] = defined( 'ACF_CODIFIER_FIELD_VALUE' ) ?
+            ACF_CODIFIER_FIELD_VALUE : 'acf_codifier/field_value';
+
+        /**
+         * This ttl is the default ttl for object caching in get_fields() method.
+         *
+         * This is overridable with the 'ACF_CODIFIER_CACHE_TTL' constant.
+         *
+         * @var float|int Defaults to 15 minutes in seconds.
+         */
+        self::$settings['cache_ttl'] = defined( 'ACF_CODIFIER_CACHE_TTL' ) ?
+            ACF_CODIFIER_CACHE_TTL : 15 * MINUTE_IN_SECONDS;
+
+    }
+
+    /**
+     * Get a plugin setting by a passed key.
+     *
+     * @param string $key The setting key.
+     *
+     * @return mixed|null The found setting or null value.
+     */
+    public static function get_setting( $key ) {
+        if ( isset( self::$settings[ $key ] ) ) {
+            return self::$settings[ $key ];
         }
 
-        if ( defined( 'ACF_CODIFIER_CACHE_TTL' ) ) {
-            self::$cache_group = ACF_CODIFIER_CACHE_TTL;
-        }
-
+        return null;
     }
 
     /**
@@ -146,12 +180,24 @@ class Codifier {
     /**
      * A more efficient replacement for ACF's native get_fields function
      *
-     * @param int $id       Post id to fetch fields from
-     * @param array $wanted Keys for wanted fields. Empty or null default to all
+     * @param int   $id     Post id to fetch fields from. If no id is passed, we get it with get_the_ID().
+     * @param array $wanted Keys for wanted fields. Empty or null default to all.
+     *
      * @return array Fields as an associative array
      */
-    public static function get_fields( $id, $wanted = null ) {
+    public static function get_fields( $id = 0, $wanted = null ) {
         global $wpdb;
+
+        // Settings needed.
+        $sorted_meta_group  = self::$settings['sorted_meta_group'];
+        $field_object_group = self::$settings['field_object_group'];
+        $field_value_group  = self::$settings['field_value_group'];
+        $cache_ttl          = self::$settings['cache_ttl'];
+
+        // If no id is set, get the current id.
+        if ( $id === 0 ) {
+            $id = get_the_ID();
+        }
 
         // Get the raw meta data from the database
         $rows = get_post_meta( $id );
@@ -170,7 +216,7 @@ class Codifier {
 
         // Sort the meta rows so that the ones deeper in the final data tree come last
         // Return the sorted data from the cache if it's stored there
-        $sorted = wp_cache_get( 'sorted_meta_data/' . $id, self::$cache_group );
+        $sorted = wp_cache_get( 'sorted_meta_data/' . $id, $sorted_meta_group );
 
         if ( ! $sorted ) {
             $sort_cache = [];
@@ -205,7 +251,7 @@ class Codifier {
                 }
             } );
             // Store the sorted data to the cache because this is a relatively heavy operation
-            wp_cache_set( 'sorted_meta_data/' . $id, $rows, self::$cache_group );
+            wp_cache_set( 'sorted_meta_data/' . $id, $rows, $sorted_meta_group );
         }
         else {
             $rows = $sorted;
@@ -242,7 +288,7 @@ class Codifier {
             $field_key = $original[ '_' . $key ];
 
             // Check if we have a field declaration stored in cache
-            $field = wp_cache_get( 'get_field_object=' . $field_key, self::$cache_group );
+            $field = wp_cache_get( 'get_field_object=' . $field_key, $field_object_group );
 
             if ( ! $field ) {
                 $field = acf_get_local_field( $field_key );
@@ -257,7 +303,7 @@ class Codifier {
                 $field['_name'] = $field['name'];
 
                 // Store the field declaration to cache
-                wp_cache_set( 'get_field_object=' . $field_key, $field, self::$cache_group );
+                wp_cache_set( 'get_field_object=' . $field_key, $field, $field_object_group );
             }
 
             // If there have been cloned fields, we need to run a few checks
@@ -297,35 +343,35 @@ class Codifier {
             }
 
             /**
-             * Filters the cache group key before field handling.
+             * Filters the cache ttl before field handling.
              *
-             * Use this to set specific group keys for different post types for instance.
+             * Use this to set specific ttl for different post types for instance.
              *
              * @since 1.2.0
              *
-             * @param string $cache_group The default group key.
-             * @param int    $id          The post id.
-             * @param string $key         The ACF field key.
+             * @param int    $field_value_ttl The cache ttl in seconds.
+             * @param int    $id              The post id.
+             * @param string $field           The ACF field object.
              */
-            $cache_group = apply_filters( 'acf_codifier_cache_key', self::$cache_group, $id, $field['key'] );
+            $field_value_ttl = apply_filters( 'acf_codifier_cache_ttl', $cache_ttl, $id, $field );
 
             // Do field type specific things
             switch ( $field['type'] ) {
                 case 'clone':
                     // Get the cloned field's field object either from cache or from the declaration
-                    $field_object = wp_cache_get( 'get_field_object/' . $field['key'], $cache_group );
+                    $field_object = wp_cache_get( 'get_field_object/' . $field['key'], $field_object_group );
                     if ( ! $field_object ) {
                         $field_object = acf_get_local_field( $field['key'] );
-                        wp_cache_set( 'get_field_object/' . $field['key'], $field_object, $cache_group );
+                        wp_cache_set( 'get_field_object/' . $field['key'], $field_object, $field_object_group, $cache_ttl );
                     }
 
                     // Loop through cloned fields and fetch their field objects
                     // either from cache or from the declaration
                     foreach ( $field_object['clone'] as $cloned_fields ) {
-                        $cloned_field = wp_cache_get( 'get_field_object/' . $cloned_fields, $cache_group );
+                        $cloned_field = wp_cache_get( 'get_field_object/' . $cloned_fields, $field_object_group );
                         if ( ! $cloned_field ) {
                             $cloned_field = acf_get_local_field( $cloned_fields );
-                            wp_cache_set( 'get_field_object/' . $cloned_fields, $cloned_field, $cache_group );
+                            wp_cache_set( 'get_field_object/' . $cloned_fields, $cloned_field, $field_object_group, $cache_ttl );
                         }
 
                         // Store the objects for future use
@@ -360,7 +406,7 @@ class Codifier {
                     $value_node = [];
                     break;
                 default:
-                    $value_node = wp_cache_get( 'get_field/formatted/' . $id . '/' . $key, $cache_group );
+                    $value_node = wp_cache_get( 'get_field/' . $id . '/' . $key, $field_value_group );
 
                     if ( ! $value_node ) {
 
@@ -374,10 +420,10 @@ class Codifier {
 
                         // Store the formatted value. This is stored for a shorter time to prevent data incoherent data.
                         wp_cache_set(
-                            'get_field/formatted/' . $id . '/' . $key,
+                            'get_field/' . $id . '/' . $key,
                             $value_node,
-                            $cache_group,
-                            self::$cache_ttl
+                            $field_value_group,
+                            $field_value_ttl
                         );
                     }
                     break;
@@ -393,8 +439,8 @@ class Codifier {
         self::ksort_recursive( $fields );
 
         // Run DustPress Components filters for the layouts through previously stored references
-        foreach( $layout_filters as $key => &$datas ) {
-            foreach( $datas as &$data ) {
+        foreach ( $layout_filters as $key => &$datas ) {
+            foreach ( $datas as &$data ) {
                 $data = apply_filters( 'dustpress/components/data=' . $key, $data );
             }
         }
@@ -405,8 +451,8 @@ class Codifier {
     /**
      * A helper method to sort arrays recursively by key
      *
-     * @param array $array       An array to sort
-     * @param int   $sort_flags  Possible sorting flags
+     * @param array $array       An array to sort.
+     * @param int   $sort_flags  Possible sorting flags.
      * @return boolean
      */
     public static function ksort_recursive( &$array, $sort_flags = SORT_REGULAR ) {
@@ -423,25 +469,45 @@ class Codifier {
     }
 
     /**
-     * Clears the cache for sorted meta rows.
-     * This is done after any meta value is changed.
+     * Controls cache flushing when a post meta value is changed.
+     * Flushes the sorted meta and the field value cache.
      *
-     * @param int $post_id Post ID.
+     * @since 1.2.0
+     *
+     * @param int    $meta_id    ID of updated metadata entry.
+     * @param int    $object_id  Object ID.
+     * @param string $meta_key   Meta key.
+     *
+     * @return null
      */
-    public static function flush_sorted_meta( $post_id ) {
-        wp_cache_delete( 'sorted_meta_data/' . $post_id, self::$cache_group );
+    public static function flush_meta_cache( $meta_id, $object_id, $meta_key ) {
+
+        /**
+         * Skip the edit lock meta key which controls post locking.
+         *
+         * @link https://codex.wordpress.org/Post_Locking
+         */
+        if ( $meta_key === '_edit_lock' ) {
+            return;
+        }
+
+        wp_cache_delete( 'sorted_meta_data/' . $object_id, self::$settings['sorted_meta_group'] );
+
+        self::flush_field_value( $object_id, $meta_key );
     }
 
     /**
-     * Flus a single value cache if its value is changed.
+     * Flushes the cache of a single field value if its meta value is updated.
+     *
+     * Note that a meta key might not match to a ACF field value. In this case
+     * WP object cache will not find a matching cache key and nothing is done.
      *
      * @link https://www.advancedcustomfields.com/resources/acf-update_value/
      *
-     * @param mixed  $value
-     * @param int    $post_id
-     * @param object $field
+     * @param int   $id       The object id.
+     * @param array $meta_key The post meta key of the field value.
      */
-    public static function flush_formatted_value( $value, $post_id, $field ) {
-
+    public static function flush_field_value( $id, $meta_key ) {
+        wp_cache_delete( 'get_field/' . $id . '/' . $meta_key, self::$settings['field_value_group'] );
     }
 }